Compute theroretical Wavelet Variance

Displays a plot of wavelet variances accounting for CI values.

Compute theroretical Wavelet Variance

# S3 method for mimu
plot(model, tau)

wv_theo(model, tau)

Arguments

model
A model object.

tau
A vector of \(2^1:number of scales\)

model
A model object.

tau
A vector of \(2^1:number of scales\)

Value

vector of theoretical Wavelet Variance for a given model

vector of theoretical Wavelet Variance for a given model

Examples

#' Xt = rnorm(n/4)
Yt = rnorm(n/2) + cumsum(rnorm(n/2, 0, 10^(-2)))
Zt = rnorm(n) + cumsum(rnorm(n, 0, 10^(-3)))
obj = make_mimu(Xt, Yt, Zt, freq = 100, unit = "s",
sensor.name = "MTiG - Gyro. X", exp.name = c("today", "yesterday", "a few days ago"))
model = AR1() + WN()

Extract the scales of first replicates of mimu \code{object} and compute tau
tau = mimu[[1]]$tau
wv_theo = function(model, tau)
#' Xt = rnorm(n/4)
Yt = rnorm(n/2) + cumsum(rnorm(n/2, 0, 10^(-2)))
Zt = rnorm(n) + cumsum(rnorm(n, 0, 10^(-3)))
obj = make_wvar_mimu_obj(Xt, Yt, Zt, freq = 100, unit = "s",
sensor.name = "MTiG - Gyro. X", exp.name = c("today", "yesterday", "a few days ago"))
model = AR1() + WN()

Extract the scales of first replicates of mimu \code{object} and compute tau
tau = 2^(1:length(sensor.name[[1]]$scales))
wv_theo = function(model, tau)

#> Error: <text>:8:9: unexpected symbol
#> 7: 
#> 8: Extract the
#>            ^